fixup! ota_update: upgrade ota with ota_update module
